% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mergeCANSIM.r
\name{mergeCANSIM}
\alias{mergeCANSIM}
\title{mergeCANSIM}
\usage{
mergeCANSIM(cansimTableVector, keepUnmatched = FALSE, showLabels = TRUE)
}
\arguments{
\item{cansimTableVector}{- vector containing all the CANSIM tables to extract and merge.}

\item{keepUnmatched}{- if true, keeps all values, even if no match was found. FALSE by default.}

\item{showLabels}{- show the Statistics Canada labels after finishing extracting and converting the table, TRUE by default.}
}
\value{
data frame containing CANSIM tables.
}
\description{
Extracts more than one CANSIM table and produces a merged table (by ID and period).
}
\examples{
mergeCANSIM( c(23100067, 23100238) )
}
