% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/searchLabels.r
\name{searchLabels}
\alias{searchLabels}
\title{searchLabels}
\usage{
  searchLabels(df, pattern)
}
\arguments{
  \item{df}{- data.frame obtained using the getCANSIM function.}
  
  \item{pattern}{- the string to search for in the variable labels.}
}
\description{
  Helper function to search through the Hmisc labels in order to obtain the variable name mapping to their respective labels.
  It can be particularly useful for bigger datasets for which manual exploration and visualization can be prohibitive.
}
\details{
  Search is case insensitive.
}
\examples{
  df <- getCANSIM("12-10-0005")
  searchLabels(df, "imports")
}
